+++ assets/plugins/wwsu-calendar/js/wwsu-calendar.js
@@ -691,6 +691,9 @@ class CalendarDb {
 		let vschedule = new WWSUdb(TAFFY());
 		vschedule.query(this.schedule.find(), true);
 
+		// Unfiltered events
+		let unfilteredEvents = [];
+
 		// prepare start and end detection
 		let start = null;
 		let end = null;
@@ -1118,8 +1121,10 @@ class CalendarDb {
 		const processEvent = (events, event, index) => {
 			// If this schedule was created as an override, we need to check to see if the override is still valid
 			if (event.overriddenID) {
-				// Find the original event
-				let record = events.find((eventb) => eventb.ID === event.overriddenID);
+				// Find the original event via the unfiltered events
+				let record = unfilteredEvents.find(
+					(eventb) => eventb.ID === event.overriddenID
+				);
 
 				// If we could not find it, the override is invalid, so we can remove it and not continue beyond this point for the event.
 				if (!record) {
@@ -1224,6 +1229,7 @@ class CalendarDb {
 			 */
 			const eventsCall = (events) => {
 				progressCallback(`Stage 3 of 4: Intelligently filtering events`);
+				unfilteredEvents = _.cloneDeep(events); // Set unfiltered events to the variable; used for some conflict checks
 				tasks = events.length;
 				tasksCompleted = 0;
 				let filteredEvents = [];